+++ src/AtomUI.Desktop.Controls/Steps/StepsItemIndicator.cs
@@ -1,11 +1,10 @@
-using AtomUI.Animations;
-using AtomUI.Controls.Primitives;
 using AtomUI.Controls;
+using AtomUI.Controls.Primitives;
 using AtomUI.Media;
+using AtomUI.Utils;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.Primitives;
-using Avalonia.Interactivity;
 using Avalonia.Media;
 
 namespace AtomUI.Desktop.Controls;
@@ -14,123 +13,131 @@ internal class StepsItemIndicator : TemplatedControl
 {
     #region 公共属性定义
 
-    public static readonly StyledProperty<StepsItemStatus> StatusProperty =
-        AvaloniaProperty.Register<StepsItemIndicator, StepsItemStatus>(nameof(Status), StepsItemStatus.Process);
-    
+    public static readonly StyledProperty<StepsStatus> StatusProperty =
+        AvaloniaProperty.Register<StepsItemIndicator, StepsStatus>(nameof(Status), StepsStatus.Process);
+
     public static readonly StyledProperty<SizeType> SizeTypeProperty =
         SizeTypeControlProperty.SizeTypeProperty.AddOwner<StepsItemIndicator>();
-    
+
     public static readonly StyledProperty<bool> IsMotionEnabledProperty =
         MotionAwareControlProperty.IsMotionEnabledProperty.AddOwner<StepsItemIndicator>();
-    
-    public static readonly StyledProperty<StepsItemIndicatorType> IndicatorTypeProperty =
-        StepsItem.IndicatorTypeProperty.AddOwner<StepsItemIndicator>();
+
+    public static readonly StyledProperty<StepsType> TypeProperty =
+        StepsItem.TypeProperty.AddOwner<StepsItemIndicator>();
 
     public static readonly StyledProperty<PathIcon?> IconProperty =
         StepsItem.IconProperty.AddOwner<StepsItemIndicator>();
-    
-    public static readonly StyledProperty<StepsStyle> StyleProperty =
-        Steps.StyleProperty.AddOwner<StepsItemIndicator>();
-    
-    public StepsItemStatus Status
+
+    public StepsStatus Status
     {
         get => GetValue(StatusProperty);
         set => SetValue(StatusProperty, value);
     }
-    
+
     public SizeType SizeType
     {
         get => GetValue(SizeTypeProperty);
         set => SetValue(SizeTypeProperty, value);
     }
-    
+
     public bool IsMotionEnabled
     {
         get => GetValue(IsMotionEnabledProperty);
         set => SetValue(IsMotionEnabledProperty, value);
     }
-    
-    public StepsItemIndicatorType IndicatorType
+
+    public StepsType Type
     {
-        get => GetValue(IndicatorTypeProperty);
-        set => SetValue(IndicatorTypeProperty, value);
+        get => GetValue(TypeProperty);
+        set => SetValue(TypeProperty, value);
     }
-    
+
     public PathIcon? Icon
     {
         get => GetValue(IconProperty);
         set => SetValue(IconProperty, value);
     }
-    
-    public StepsStyle Style
-    {
-        get => GetValue(StyleProperty);
-        set => SetValue(StyleProperty, value);
-    }
 
     #endregion
-    
+
     #region 内部属性定义
 
-    internal static readonly DirectProperty<StepsItemIndicator, int> PositionProperty =
+    internal static readonly DirectProperty<StepsItemIndicator, int> StepNumberProperty =
         AvaloniaProperty.RegisterDirect<StepsItemIndicator, int>(
-            nameof(Position),
-            o => o.Position,
-            (o, v) => o.Position = v);
-    
+            nameof(StepNumber),
+            indicator => indicator.StepNumber,
+            (indicator, value) => indicator.StepNumber = value);
+
+    internal static readonly DirectProperty<StepsItemIndicator, int> DisplayStepNumberProperty =
+        AvaloniaProperty.RegisterDirect<StepsItemIndicator, int>(
+            nameof(DisplayStepNumber),
+            indicator => indicator.DisplayStepNumber);
+
     internal static readonly DirectProperty<StepsItemIndicator, bool> IsCustomProperty =
         AvaloniaProperty.RegisterDirect<StepsItemIndicator, bool>(
             nameof(IsCustom),
-            o => o.IsCustom,
-            (o, v) => o.IsCustom = v);
-    
+            indicator => indicator.IsCustom,
+            (indicator, value) => indicator.IsCustom = value);
+
     internal static readonly DirectProperty<StepsItemIndicator, bool> IsCurrentProperty =
         AvaloniaProperty.RegisterDirect<StepsItemIndicator, bool>(
             nameof(IsCurrent),
-            o => o.IsCurrent,
-            (o, v) => o.IsCurrent = v);
-    
+            indicator => indicator.IsCurrent,
+            (indicator, value) => indicator.IsCurrent = value);
+
     internal static readonly DirectProperty<StepsItemIndicator, bool> IsClickableProperty =
         AvaloniaProperty.RegisterDirect<StepsItemIndicator, bool>(
             nameof(IsClickable),
-            o => o.IsClickable,
-            (o, v) => o.IsClickable = v);
-    
+            indicator => indicator.IsClickable,
+            (indicator, value) => indicator.IsClickable = value);
+
     internal static readonly DirectProperty<StepsItemIndicator, bool> IsItemHoverProperty =
         AvaloniaProperty.RegisterDirect<StepsItemIndicator, bool>(
             nameof(IsItemHover),
-            o => o.IsItemHover,
-            (o, v) => o.IsItemHover = v);
-    
-    internal static readonly DirectProperty<StepsItemIndicator, bool> IsShowProgressProperty =
+            indicator => indicator.IsItemHover,
+            (indicator, value) => indicator.IsItemHover = value);
+
+    internal static readonly DirectProperty<StepsItemIndicator, bool> IsProgressVisibleProperty =
         AvaloniaProperty.RegisterDirect<StepsItemIndicator, bool>(
-            nameof(IsShowProgress),
-            o => o.IsShowProgress,
-            (o, v) => o.IsShowProgress = v);
-    
-    internal static readonly DirectProperty<StepsItemIndicator, double> ProgressValueProperty =
-        AvaloniaProperty.RegisterDirect<StepsItemIndicator, double>(
-            nameof(ProgressValue),
-            o => o.ProgressValue,
-            (o, v) => o.ProgressValue = v);
-    
+            nameof(IsProgressVisible),
+            indicator => indicator.IsProgressVisible,
+            (indicator, value) => indicator.IsProgressVisible = value);
+
+    internal static readonly DirectProperty<StepsItemIndicator, double?> PercentProperty =
+        AvaloniaProperty.RegisterDirect<StepsItemIndicator, double?>(
+            nameof(Percent),
+            indicator => indicator.Percent,
+            (indicator, value) => indicator.Percent = value);
+
     internal static readonly StyledProperty<double> ProgressLineThicknessProperty =
         AvaloniaProperty.Register<StepsItemIndicator, double>(nameof(ProgressLineThickness));
-    
+
     internal static readonly StyledProperty<IBrush?> ProgressGrooveColorProperty =
         AvaloniaProperty.Register<StepsItemIndicator, IBrush?>(nameof(ProgressGrooveColor));
 
     internal static readonly StyledProperty<IBrush?> ProgressColorProperty =
         AvaloniaProperty.Register<StepsItemIndicator, IBrush?>(nameof(ProgressColor));
-    
-    private int _position;
 
-    internal int Position
+    private int _stepNumber;
+
+    internal int StepNumber
     {
-        get => _position;
-        set => SetAndRaise(PositionProperty, ref _position, value);
+        get => _stepNumber;
+        set
+        {
+            SetAndRaise(StepNumberProperty, ref _stepNumber, value);
+            DisplayStepNumber = value + 1;
+        }
     }
-    
+
+    private int _displayStepNumber = 1;
+
+    internal int DisplayStepNumber
+    {
+        get => _displayStepNumber;
+        private set => SetAndRaise(DisplayStepNumberProperty, ref _displayStepNumber, value);
+    }
+
     private bool _isCustom;
 
     internal bool IsCustom
@@ -138,7 +145,7 @@ internal class StepsItemIndicator : TemplatedControl
         get => _isCustom;
         set => SetAndRaise(IsCustomProperty, ref _isCustom, value);
     }
-    
+
     private bool _isCurrent;
 
     internal bool IsCurrent
@@ -146,7 +153,7 @@ internal class StepsItemIndicator : TemplatedControl
         get => _isCurrent;
         set => SetAndRaise(IsCurrentProperty, ref _isCurrent, value);
     }
-    
+
     private bool _isClickable;
 
     internal bool IsClickable
@@ -154,7 +161,7 @@ internal class StepsItemIndicator : TemplatedControl
         get => _isClickable;
         set => SetAndRaise(IsClickableProperty, ref _isClickable, value);
     }
-    
+
     private bool _isItemHover;
 
     internal bool IsItemHover
@@ -162,51 +169,76 @@ internal class StepsItemIndicator : TemplatedControl
         get => _isItemHover;
         set => SetAndRaise(IsItemHoverProperty, ref _isItemHover, value);
     }
-    
-    private bool _isShowProgress;
 
-    internal bool IsShowProgress
-    {
-        get => _isShowProgress;
-        set => SetAndRaise(IsShowProgressProperty, ref _isShowProgress, value);
-    }
-    
-    private double _progressValue;
+    private bool _isProgressVisible;
 
-    internal double ProgressValue
+    internal bool IsProgressVisible
     {
-        get => _progressValue;
-        set => SetAndRaise(ProgressValueProperty, ref _progressValue, value);
+        get => _isProgressVisible;
+        set => SetAndRaise(IsProgressVisibleProperty, ref _isProgressVisible, value);
     }
-    
+
+    private double? _percent;
+
+    internal double? Percent
+    {
+        get => _percent;
+        set => SetAndRaise(PercentProperty, ref _percent, value);
+    }
+
     internal double ProgressLineThickness
     {
         get => GetValue(ProgressLineThicknessProperty);
         set => SetValue(ProgressLineThicknessProperty, value);
     }
-    
+
     internal IBrush? ProgressGrooveColor
     {
         get => GetValue(ProgressGrooveColorProperty);
         set => SetValue(ProgressGrooveColorProperty, value);
     }
-    
+
     internal IBrush? ProgressColor
     {
         get => GetValue(ProgressColorProperty);
         set => SetValue(ProgressColorProperty, value);
     }
+
     #endregion
 
     private WaveSpiritDecorator? _waveSpiritDecorator;
+    private IPen? _progressGroovePen;
+    private IPen? _progressPen;
+
+    internal bool IsWavePlaying => _waveSpiritDecorator?.IsPlaying == true;
 
     static StepsItemIndicator()
     {
-        AffectsMeasure<StepsItemIndicator>(SizeTypeProperty, IndicatorTypeProperty, IsCurrentProperty,
-            IsShowProgressProperty,
-            StyleProperty);
+        AffectsMeasure<StepsItemIndicator>(
+            SizeTypeProperty,
+            TypeProperty,
+            IsCurrentProperty,
+            IsProgressVisibleProperty);
+        AffectsRender<StepsItemIndicator>(
+            PercentProperty,
+            ProgressLineThicknessProperty,
+            ProgressGrooveColorProperty,
+            ProgressColorProperty,
+            IsCurrentProperty,
+            IsProgressVisibleProperty,
+            UseLayoutRoundingProperty);
     }
-    
+
+    internal void PlayWave()
+    {
+        if (!IsEnabled || !IsMotionEnabled || !IsLoaded || _waveSpiritDecorator is null)
+        {
+            return;
+        }
+
+        _waveSpiritDecorator.Play();
+    }
+
     protected override void OnSizeChanged(SizeChangedEventArgs e)
     {
         base.OnSizeChanged(e);
@@ -218,63 +250,36 @@ internal class StepsItemIndicator : TemplatedControl
         base.OnPropertyChanged(change);
         if (change.Property == IconProperty)
         {
-            SetCurrentValue(IsCustomProperty, Icon != null);
-        }
-        else if (change.Property == IsCurrentProperty && change.GetNewValue<bool>())
-        {
-            PlayWave();
+            SetCurrentValue(IsCustomProperty, Icon is not null);
         }
     }
 
     protected override void OnApplyTemplate(TemplateAppliedEventArgs e)
     {
+        _waveSpiritDecorator = null;
         base.OnApplyTemplate(e);
         _waveSpiritDecorator = e.NameScope.Find<WaveSpiritDecorator>(WaveSpiritDecorator.WaveSpiritPart);
-    }
-
-    private void PlayWave()
-    {
-        if (!IsClickable || !IsEnabled || !IsMotionEnabled || !IsLoaded || _waveSpiritDecorator is null)
-        {
-            return;
-        }
-
-        _waveSpiritDecorator.Play();
     }
 
     public override void Render(DrawingContext context)
     {
         base.Render(context);
-        if (IsShowProgress && IsCurrent)
+        if (!IsProgressVisible || !Percent.HasValue)
         {
-            var progressRect = new Rect(DesiredSize.Deflate(Margin)).Deflate(ProgressLineThickness / 2);
-            {
-                var pen          = new Pen(ProgressGrooveColor, ProgressLineThickness);
-                context.DrawEllipse(null, pen, progressRect);
-            }
-
-            {
-                var pen = new Pen(ProgressColor, ProgressLineThickness)
-                {
-                    LineCap = PenLineCap.Round
-                };
-
-                double startAngle = -90;
-                var indicatorAngle = 360 * ProgressValue / 100;
-                context.DrawArc(pen, progressRect, startAngle, indicatorAngle);
-            }
+            return;
         }
-    }
 
-    protected override void OnInitialized()
-    {
-        base.OnInitialized();
-        this.DisableTransitions();
-    }
+        var lineThickness = BorderUtils.BuildRenderScaleAwareThickness(this, ProgressLineThickness);
+        var progressRect = new Rect(Bounds.Size).Deflate(lineThickness / 2);
 
-    protected override void OnLoaded(RoutedEventArgs e)
-    {
-        base.OnLoaded(e);
-        Dispatcher.Post(this.EnableTransitions);
+        PenUtils.TryModifyOrCreate(ref _progressGroovePen, ProgressGrooveColor, lineThickness);
+        context.DrawEllipse(null, _progressGroovePen, progressRect);
+
+        PenUtils.TryModifyOrCreate(
+            ref _progressPen,
+            ProgressColor,
+            lineThickness,
+            lineCap: PenLineCap.Round);
+        context.DrawArc(_progressPen, progressRect, -90, 360 * Percent.Value / 100);
     }
 }